+++ hosts/phi/configuration.nix
@@ -55,6 +55,7 @@
   users.users.${user} = {
     openssh.authorizedKeys.keys = builtins.attrValues {
       inherit (keys.users) enzime_sigma;
+      inherit (keys.hosts) sigma;
     };
   };