+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `/mnt/phi` automount on `sigma`
 - Fixed `slack` on `x86_64-darwin`
 - Use `x86_64-darwin` on `psi`
 - Use 1Password SSH agent on `sigma`